feat: swap coins-bip32/39 for RustCrypto bip32/bip39 (#12)
* feat: swap coins-bip32/39 for RustCrypto bip32/bip39 (#11)

The summa-tx coins crates hard-pinned base64 = "0.21" via coins-core
and the upstream repo has been quiet for the last year, so the
dedup against the rest of the wallet's base64 0.22 chain was
blocked. Switching to RustCrypto's bip32 0.5 + bip39 2 unifies the
keycard stack on a single crypto-crate vendor (which we already
pull via k256 / elliptic-curve) and drops base64 0.21 + the three
coins-* crates from the dependency tree.

Source impact is small. `bip32::DerivationPath` and
`bip32::ChildNumber` cover what we used (parse / iter / push /
Display); ChildNumber's `From<u32>` + `From<ChildNumber> for u32`
make the BIP-32 wire-format conversion in
`derivation_path_to_bytes` / `get_status` / `set_pinless_path`
one-liners. `bip39::Mnemonic` is non-generic over language so
`generate_mnemonic(words)` becomes
`generate_mnemonic(words, Language)`; CLI commands collapse from
the old generic-over-Wordlist helper into a runtime match.

Closes #11.

* style: rustfmt import order + line breaks

+++ nexum-keycard/src/commands/mod.rs
@@ -1,6 +1,6 @@
 pub mod derive_key;
+use bip32::DerivationPath;
 use bytes::{Bytes, BytesMut};
-use coins_bip32::path::DerivationPath;
 pub use derive_key::*;
 pub mod export_key;
 pub use export_key::*;
@@ -60,7 +60,7 @@ pub enum PersistentRecord {
 pub(crate) fn derivation_path_to_bytes(path: &DerivationPath) -> Bytes {
     path.iter()
         .fold(BytesMut::new(), |mut bytes, component| {
-            bytes.extend_from_slice(&component.to_be_bytes());
+            bytes.extend_from_slice(&u32::from(component).to_be_bytes());
             bytes
         })
         .freeze()
